feat: implement locale middleware for language detection and message handling
